refactor : domain, view 코드 분리
