+++ week5-assingment/index.html
@@ -9,7 +9,8 @@
       href="https://fonts.googleapis.com/css2?family=Material+Symbols+Outlined:opsz,wght,FILL,GRAD@24,400,0,0"
     />
     <link rel="stylesheet" href="./style.css" />
-    <script defer src="./index.js"></script>
+    <script type="module" defer src="./domain.js"></script>
+    <script type="module" defer src="./view.js"></script>
   </head>
 
   <body>
